split backend entry point
