+++ packages/server/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "version": "1.0.0",
   "scripts": {
-    "dev": "node --watch ./src/index.js",
-    "start": "node ./src/index.js",
+    "dev": "node --watch ./src/server.js",
+    "start": "node ./src/server.js",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage"
